Move translation of visual metadata to mpl
